requirements.txt
azure-core==1.17.0
azureml-core==1.33.0
azureml-dataprep==2.20.1
azureml-pipeline==1.33.0
azureml-pipeline-core==1.33.0
azureml-pipeline-steps==1.33.0
pandas==1.1.2